src/requirements.txt
fastapi==0.111.0
uvicorn==0.27.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib